Add UserAuth class with verifyUserOwnership() to UserController methods

+++ src/controller/userController.js
@@ -1,23 +1,28 @@
 import User from '../models/User.js'
 
 export class UserController {
-   
+    //  Technically no longer needed due to authController.register() handling this...
     static async create(req, res) {
-        console.log('Cuerpo de la solicitud:', req.body) 
         const { name_user, email, password_user } = req.body
+        
         try {
             const userId = await User.createUser(name_user, email, password_user)
+            
             res.status(201).json({ id: userId })
         } catch (error) {
-            console.error('Error creando usuario:', error)
-            res.status(500).json({ error: error.message || 'Error creando usuario' })
+            res.status(500).json({ error: 'Error creating user' })
         }
     }
     
     static async get(req, res) {
         const { id } = req.params
+        const { id: USER_ID } = req.user
+
         try {
+            await UserAuth.verifyUserOwnership(id, USER_ID)
+            
             const user = await User.getUserById(id)
+            
             if (user) {
                 res.json(user)
             } else {
@@ -31,15 +36,18 @@ export class UserController {
     static async update(req, res) {
         const { id } = req.params
         const { name_user, email, password_user} = req.body
-    
-        // Crear un objeto con los campos opcionales que se quieren actualizar
+        const { id: USER_ID } = req.user
         const data = {}
-        if (name_user) data.name_user = name_user
-        if (email) data.email = email
-        if (password_user) data.password_user = password_user
     
         try {
+            await UserAuth.verifyUserOwnership(id, USER_ID)
+
+            if (name_user) data.name_user = name_user
+            if (email) data.email = email
+            if (password_user) data.password_user = password_user
+
             await User.updateUser(id, data)
+
             res.json({ message: 'User updated successfully' })
         } catch (error) {
             res.status(500).json({ error: 'Error updating user' })
@@ -48,11 +56,26 @@ export class UserController {
     
     static async delete(req, res) {
         const { id } = req.params
+        const { id: USER_ID } = req.user
+
         try {
+            await UserAuth.verifyUserOwnership(id, USER_ID)
+
             await User.deleteUser(id)
+
             res.json({ message: 'User deleted successfully' })
         } catch (error) {
             res.status(500).json({ error: 'Error deleting user' })
         }
+    }
+}
+
+export class UserAuth {
+    static async verifyUserOwnership(target_id, USER_ID) {
+        const TARGETED_USER = await Task.getUserById(target_id)
+            
+        if (!TARGETED_USER) throw new Error('User not found')
+
+        if (target_id !== USER_ID) throw new Error('Invalid user ID')
     }
 }
